\name{AnomalyScore}
\alias{AnomalyScore}

\title{Evaluating isolation trees}
\description{
Return anomaly scores and path length of \code{x} according to \code{forest}.
 }
\usage{
       AnomalyScore(x, forest, ntree = forest$ntree, hlim = forest$hlim, appRange = F)
}
\arguments{
  \item{x}{a data frame of samples}
  \item{forest}{an object generated by \code{IsolationTrees()}}
  \item{ntree}{specifies the number of trees used to find the anomaly score, must be greater than 0 and smaller or equal to number of tree used in the model}
  \item{hlim}{evaluation tree height limit}
  \item{appRange}{boolean switch to apply application range}
}
\value{
\item{outF}{anomaly scores of samples in \code{x}}
\item{pathLength}{average path lengths of samples in \code{x}}
}
\details{
 \code{AnomalyScore} passes each sample in \code{x} through \code{forest} to obtain expected path length.  It then uses path length and parameter in \code{forest} to determine anomaly score.
 }
\references{Fei Tony Liu, Kai Ming Ting, and Zhi-Hua Zhou\cr
 \emph{Isolation Forest}\cr 
 IEEE International Conference on Data Mining 2008 (ICDM 08), Pisa, Italy, 2008.
   \url{http://www.gscit.monash.edu.au/gscitweb/loid.php?loid=905282&mimetype=application/pdf}
}
\seealso{
  \code{\link{IsolationTrees}}
}
\examples{

library(IsolationForest)
data(stackloss)
# train a model of Isolation Forest
tr<-IsolationTrees(stackloss,rFactor=0)
#evaluate anomaly score
as<-AnomalyScore(stackloss,tr)
# show anomaly score
as$outF

}
\author{Fei Tony Liu}
\keyword{models}


